+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/execution/versions/eip4844/BlobsSidecar.java
@@ -17,54 +17,17 @@ import java.util.List;
 import org.apache.tuweni.bytes.Bytes32;
 import tech.pegasys.teku.infrastructure.ssz.SszList;
 import tech.pegasys.teku.infrastructure.ssz.containers.Container4;
-import tech.pegasys.teku.infrastructure.ssz.containers.ContainerSchema4;
 import tech.pegasys.teku.infrastructure.ssz.primitive.SszBytes32;
 import tech.pegasys.teku.infrastructure.ssz.primitive.SszUInt64;
-import tech.pegasys.teku.infrastructure.ssz.schema.SszFieldName;
-import tech.pegasys.teku.infrastructure.ssz.schema.SszListSchema;
-import tech.pegasys.teku.infrastructure.ssz.schema.SszPrimitiveSchemas;
 import tech.pegasys.teku.infrastructure.ssz.tree.TreeNode;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.kzg.KZGProof;
-import tech.pegasys.teku.spec.config.SpecConfigEip4844;
 import tech.pegasys.teku.spec.datastructures.type.SszKZGProof;
-import tech.pegasys.teku.spec.datastructures.type.SszKZGProofSchema;
 
 public class BlobsSidecar
     extends Container4<BlobsSidecar, SszBytes32, SszUInt64, SszList<Blob>, SszKZGProof> {
 
-  public static class BlobsSidecarSchema
-      extends ContainerSchema4<BlobsSidecar, SszBytes32, SszUInt64, SszList<Blob>, SszKZGProof> {
-
-    static final SszFieldName FIELD_BLOBS = () -> "blobs";
-
-    BlobsSidecarSchema(final SpecConfigEip4844 specConfig, final BlobSchema blobSchema) {
-      super(
-          "BlobsSidecar",
-          namedSchema("beacon_block_root", SszPrimitiveSchemas.BYTES32_SCHEMA),
-          namedSchema("beacon_block_slot", SszPrimitiveSchemas.UINT64_SCHEMA),
-          namedSchema(
-              FIELD_BLOBS, SszListSchema.create(blobSchema, specConfig.getMaxBlobsPerBlock())),
-          namedSchema("kzg_aggregated_proof", SszKZGProofSchema.INSTANCE));
-    }
-
-    @SuppressWarnings("unchecked")
-    public SszListSchema<Blob, ?> getBlobsSchema() {
-      return (SszListSchema<Blob, ?>) getChildSchema(getFieldIndex(FIELD_BLOBS));
-    }
-
-    @Override
-    public BlobsSidecar createFromBackingNode(final TreeNode node) {
-      return new BlobsSidecar(this, node);
-    }
-  }
-
-  public static BlobsSidecarSchema create(
-      final SpecConfigEip4844 specConfig, final BlobSchema blobSchema) {
-    return new BlobsSidecarSchema(specConfig, blobSchema);
-  }
-
-  private BlobsSidecar(final BlobsSidecarSchema type, final TreeNode backingNode) {
+  BlobsSidecar(final BlobsSidecarSchema type, final TreeNode backingNode) {
     super(type, backingNode);
   }
